+++ exports/index.js
@@ -1,5 +1,5 @@
-export { default as LightCode } from "./light-code-register.js";
-export { default as LightDisclosure } from "./light-disclosure-register.js";
-export { default as LightEditor } from "./light-editor-register.js";
-export { default as LightPen } from "./light-pen-register.js";
-export { default as LightPreview } from "./light-preview-register.js";
+export { default as LightCode } from "./components/light-code/light-code-register.js";
+export { default as LightDisclosure } from "./components/light-disclosure/light-disclosure-register.js";
+export { default as LightEditor } from "./components/light-editor/light-editor-register.js";
+export { default as LightPen } from "./components/light-pen/light-pen-register.js";
+export { default as LightPreview } from "./components/light-preview/light-preview-register.js";